feat: 优化收到命令提醒
