+++ wechatter/sender/__init__.py
@@ -1,3 +1,4 @@
+from .notifier import notify_logged_in, notify_logged_out, notify_received
 from .sender import Sender
 
-__all__ = ["Sender"]
+__all__ = ["Sender", "notify_received", "notify_logged_in", "notify_logged_out"]
